Refactoring/moving to separate packages
