+++ storage/test/integration/variant/service/conftest.py
@@ -13,10 +13,10 @@ from storage.variant.service import DatabaseConnection
 from storage.variant.service.ReferenceService import ReferenceService
 from storage.variant.service.VariationService import VariationService
 from storage.variant.service.CoreAlignmentService import CoreAlignmentService
-from storage.variant.io.SnippyVariantsReader import SnippyVariantsReader
-from storage.variant.io.VcfVariantsReader import VcfVariantsReader
-from storage.variant.io.MLSTFeaturesReader import MLSTFeaturesReader
-from storage.variant.io.BasicMLSTFeaturesReader import BasicMLSTFeaturesReader
+from storage.variant.io.mutation.SnippyVariantsReader import SnippyVariantsReader
+from storage.variant.io.mutation.VcfVariantsReader import VcfVariantsReader
+from storage.variant.io.mlst.MLSTFeaturesReader import MLSTFeaturesReader
+from storage.variant.io.mlst.MLSTTSeemannFeaturesReader import MLSTTSeemannFeaturesReader
 from storage.variant.service.SampleService import SampleService
 from storage.variant.service.TreeService import TreeService
 from storage.variant.service.KmerQueryService import KmerQueryService
@@ -152,12 +152,12 @@ def kmer_query_service_with_data(sample_service, kmer_service_with_data) -> Kmer
 
 @pytest.fixture
 def mlst_reader_single_scheme() -> MLSTFeaturesReader:
-    return BasicMLSTFeaturesReader(mlst_file=mlst_file_single_scheme)
+    return MLSTTSeemannFeaturesReader(mlst_file=mlst_file_single_scheme)
 
 
 @pytest.fixture
 def mlst_reader_basic() -> MLSTFeaturesReader:
-    return BasicMLSTFeaturesReader(mlst_file=basic_mlst_file)
+    return MLSTTSeemannFeaturesReader(mlst_file=basic_mlst_file)
 
 
 @pytest.fixture
