+++ storage/variant/service/CoreAlignmentService.py
@@ -11,7 +11,7 @@ from Bio.Seq import Seq
 from Bio.SeqRecord import SeqRecord
 
 from storage.variant.MaskedGenomicRegions import MaskedGenomicRegions
-from storage.variant.io.VariationFile import VariationFile
+from storage.variant.io.mutation.VariationFile import VariationFile
 from storage.variant.model import SampleNucleotideVariation
 from storage.variant.service import DatabaseConnection
 from storage.variant.service.ReferenceService import ReferenceService
